Первый коммит. Архив                                                                                 По книге шилда                                                                                 Разбираем Свичь!!!!                                                                                 Архивный коммит                                                                                 Переделал угадайки в jar для запуска в bat нике                                                                                 угадайка символов в клаве
                                                                                Начало прграммы рандом v.00003

                                                                                Новые методы в классе Car
                                                                                Контрольный коммит+пуш
                                                                                Добавлен класс для работы с методом finalize
                                                                                Разбираем массивы  v_0.2
                                                                                Разбираем пузерьковую сортировку
                                                                                Разбор сортировки
                                                                                Разбираем и  реализовываем класс на целендричискую очередь
                                                                                Сортировка массивов с помощью встроеннного Класса Math

+++ src/com/ru/Random/Voda/com/com/ru/Zadachki/Book.java
@@ -9,14 +9,14 @@ public class Book {
    private String avtor = "Не задано";
    private int colichestvoStranic;
    private int stoinost;
-   private  byte kolichestvProdaj;
-   private byte godVipuska;
+   private  int kolichestvProdaj;
+   private int godVipuska;
 
-    public byte getGodVipuska() {
+    public int getGodVipuska() {
         return godVipuska;
     }
 
-    public void setGodVipuska(byte godVipuska) {
+    public void setGodVipuska(int godVipuska) {
         this.godVipuska = godVipuska;
     }
 
@@ -57,7 +57,7 @@ public class Book {
         this.stoinost = stoinost;
     }
 
-    public byte getKolichestvProdaj() {
+    public int getKolichestvProdaj() {
         return kolichestvProdaj;
     }
